Social Login,email verification and more features

+++ app/Http/Controllers/QuestionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Question;
 use App\Rules\KurdishOrArabicChars;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Validator;
 
 class QuestionController extends Controller
@@ -22,12 +23,13 @@ class QuestionController extends Controller
     {
         return Question::where('user_id', auth()->id());
     }
+
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
-        if (!auth()->id()) return response()->json(['errors' => "unauthorized"], 401);
+        $this->authorize('create', Question::class);
 
         $validator = Validator::make($request->all(), [
             'body' => ['required', 'min:10', new KurdishOrArabicChars]
@@ -57,8 +59,9 @@ class QuestionController extends Controller
      */
     public function update(Request $request, Question $question)
     {
+        $auth = Gate::inspect('update', $question);
+        if (!$auth->allowed()) return response()->json(['errors' => "Unauthorized user"], 403);
 
-        if (auth()->id() !== $question->user_id) return response()->json(['errors' => $question->user->name . " not unauthorized to edit this question"], 401);
 
         $validator = Validator::make($request->all(), [
             'body' => ['min:10', new KurdishOrArabicChars]
@@ -79,6 +82,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        $auth = Gate::inspect('delete', $question);
+        if (!$auth->allowed()) return response()->json(['errors' => "Unauthorized user"], 403);
+
+
         $question->delete();
 
         return ['success' => $question->body . " has been deleted successfully"];
